CI: fix typo causing only ignored tests to run in CI
This change fixes up a typo in the test workflow in CI that's causing only
ignored tests to run.

Confusingly, `--ignored` runs *only* ignored tests, but `--include-ignored`
runs tests whether they're ignored or not. `--include-ignored` is probably
what we want.

Even more confusingly, the `--ignored` and `--include-ignored` flags don't
seem to be well documented, but it is mentioned in The Book here:
https://doc.rust-lang.org/book/ch11-02-running-tests.html#ignoring-some-tests-unless-specifically-requested.